selenium_python/requirements.txt
selenium==4.38.0
webdriver-manager==4.0.2
pytest==7.4.3
python-dotenv==1.0.0
pytest-html==4.1.1
